requirements.txt
Django>=5.0,<6.0
python-dotenv==1.1.0
google-generativeai>=0.8.0
django-htmx==1.21.0
Pillow>=10.0
whitenoise>=6.5
